feat(Teleport): extract teleport disable on headset collision
The ability to disable the teleport on headset collision has now been
extracted into it's own script to decouple the feature and also to
make it optional in case it's valid for teleporting to happen on
headset collision.
