+++ Assets/VRTK/Scripts/VRTK_TeleportDisableOnHeadsetCollision.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3312e673b929b694d910e18143c4eb20
+timeCreated: 1473021823
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
